audio-emotion-server/requirements.txt
fastapi==0.100.0
uvicorn==0.23.1
python-multipart==0.0.6
pydub==0.25.1
numpy==1.25.1
torch==2.0.1
transformers==4.30.2
torchaudio==2.0.2
pyjwt==2.7.0